+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
+## [2.3.0] - 
+### Added
+- Add event entity for compact projection
+## [2.2.1] - 12/13/2018
+### Added
+- Add marketType field for MarketChange entity
+- Add name, group, outcomeType fields for SelectionChange entity
 ## [2.2.0] - 12/7/2018
 ### Added
 - Add sports data api contracts